Rename: 과제방 공용 컴포넌트 폴더명 변경, 경로 변경
