+++ src/app/assignment/(components)/Card.tsx
